Use full access field in user model

+++ api/views/shippy.py
@@ -59,7 +59,7 @@ class V1MaintainersChunkedUpload(ChunkedUploadView):
             )
 
         # Check if maintainer is in device's approved maintainers list
-        if self.request.user not in device.maintainers.all():
+        if not self.request.user.full_access_to_devices and self.request.user not in device.maintainers.all():
             chunked_upload.delete()
             return Response(
                 {
